services/ml-detector/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
numpy==1.26.2
scikit-learn==1.3.2
